requirements.txt
anthropic>=0.7.0
neo4j>=5.0.0
numpy>=1.20.0
openai>=1.0.0
pandas>=1.5.0
python-docx>=1.0.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0
streamlit>=1.25.0
tiktoken>=0.5.0
